build: include grommunio-postfix
grommunio-postfix exists, since we deploy the default configuration with the appliance,
however there have been some maps addditions which some installations are missing.

grommunio-update calls grommunio-postfix as part of the update to make sure the
intended amount of grommunio postfix maps exist and are in place.

Any other general update does not fit to the deployment (package update for example)
since the configuration files are dynamic with the content (mysql data).